impl documentSymbolVisitor
